Creating alternative Python algorithm for Check Is Square
